Transform Lunatik into loadable modules
Allow the the core Lua engine to be built as the loadable module
lunatik_core.ko.  This refactorization includes EXPORT_SYMBOL for all
the Lua functions mentioned in the header files.  This allows Lua to be
used from other loadable modules as well.

The bindings were split-off into four separate loadable modules:
 * lunatik_printk.ko
 * lunatik_inspect.ko
 * lunatik_crypto.ko
 * lunatik_buffer.ko

Those features can be selected through individual config variables.

The system call sys_lua, however, can only be built, if the core engine
was selected to be built in.  It is therefore available as a separate
config option.  This does not harm the ability to build feature bindings
as a module.

The Lua implementation files were moved into the parent lunatik
directory to enable the module build.  A shell script (gen_ksyms.sh) was
added that extracts the public functions from the Lua headers and
generates a C file (ksyms.c) that is then compiled into the module or
the kernel.  The setjmp/longjmp Implementation was finally copied from
UML to avoid various hacks needed to make it available in Lua otherwise.
UML with Lunatik enabled is currently not supported.