Remove LUAI_FUNC declarations from ksyms + cleanup
The LUAI_FUNC declarations were never meant to be publicly exposed
(i.e. are not listed in the public headers).  So there is no reason to
EXPORT_SYMBOL them.  Do not emit include directives for the headers
which do not have any publicly relevant declarations.

Also cleanup and restructure the gen_ksyms.sh script for better
readability.