Support additional return types in Lunatik
From lunatik_loadcode return a structure instead of just a char *.  This
structure contains a union of additional literal return types of Lua.
So instead of just strings (and numbers converted to strings) Lunatik
exposes with this patch the following Lua types to other parts of the
kernel.

 * LUA_TNIL (recognizable by the type field alone)
 * LUA_TNUMBER (not converted to a string)
 * LUA_TBOOLEAN
 * LUA_TSTRING (now with a separate size field and support for strings
   that contain zero byte characters)
 * LUA_TUSERDATA (as a void *, a separate size field, and an additional
   type field containing the name of the metatable)
 * LUA_TLIGHTUSERDATA

This change is especially useful because of the zero byte character
support and the explicit support for userdata.  Without this change
binary return values were impossible to handle without quoting zero byte
characters somehow in a string return value.

For exposing the name of userdata values the setup of the "__name"
metafield from Lua 5.3 was backported into luaL_newmetatable.

For the system call sys_lua the structure is transformed into three user
variables passed by reference: the type, the data (as a preallocated
userspace buffer), and the size of the buffer, while the size is also
passed by reference and the referenced userspace variable is used to
return the actual length of the returned data.  For userdata the type
name is prepended as a zero terminated string to the userdata, i.e. the
real userdata starts after the first zero byte in the data buffer.
As the buffer needs to be preallocated by userspace we have no other
chance than truncating the result if it does not fit into that buffer.
Applications need to ensure themselves to allocate a buffer big enough
for all possible (relevant) results of the code they send.