Use dedicated context for Lunatik syscall
Create a fresh context for each Lunatik syscall.  This adds isolation
between Lua executions from various places.

To allow memory cleanup in the asynchronous case (without result pointer
arguments) provide additional variants of the Lunatik loadcode
interface.  These take a callback argument and a general purpose pointer
that will be supplied to the callback which is called after completion
of the Lua execution.  The callback also receives the error code as one
argument.  There are actually two asynchronous variants: one that also
provides the result of the Lua execution to the callback and another one
that silently and efficiently ignores the result.  The syscall used the
latter variant for asynchronous executions.  For synchronous executions
a new variant is provided, too.  This variant always waits for the
execution to finish (using a semaphore), regardless of the result
pointer.  It still uses the workqueue for fairness with other executions
on the same context.  The syscall uses this variant for its synchronous
path (with result pointer arguments) for consistency reasons.

As the different loadcode behaviors are now all implemented with
explicit variants, the former lunatik_loadcode is now deprecated.  One
remarkable difference between the asynchronous path of the former
lunatik_loadcode and the new async variants is that for the new variants
the caller is expected to free the supplied code string if need be.
This can be done in the callback function.  This now allows constant
strings (that shall not be freed) be executed asynchronously or the same
code string be executed repeatedly.