Print executed Lua script completely for debug
Previously the Lua script was truncated after about 1024 characters
because the format output in printk is going to a fixed size buffer of
that length.  This patch splits the output into one printk per source
code line and also adds a line number prefix.  Lua scripts that do not
contain newline characters are continued to be printed on the same
printk line as the address.